Move SourceCodeUrl in the source_code_url module
